requirements.txt
PyQt5 >= 5.15.0
PyPDF2 >= 1.26.0
Pillow >= 9.0.0
python-docx